requirements.txt
opencv-python
torch
torchvision
facenet-pytorch
numpy
Pillow
PyYAML
matplotlib
joblib
scikit-learn
